+++ src/main/java/com/nhnacademy/frontserver1/presentation/controller/AuthContorller.java
@@ -3,11 +3,15 @@ package com.nhnacademy.frontserver1.presentation.controller;
 import com.nhnacademy.frontserver1.application.service.impl.AuthServiceImpl;
 import com.nhnacademy.frontserver1.application.service.impl.PaycoServiceImpl;
 import com.nhnacademy.frontserver1.presentation.dto.request.user.LoginUserRequest;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import com.nhnacademy.frontserver1.common.context.TokenContext;
 
 /**
  * AuthController 클래스는 사용자 인증 관련 기능을 제공하는 Spring MVC 컨트롤러입니다.
@@ -60,6 +64,17 @@ public class AuthContorller {
         authService.loginUser(loginUserRequest);
 
         return "redirect:/";
+    }
+
+    private void addTokenCookie(HttpServletResponse response, String name, String token) {
+        if (token != null) {
+            Cookie cookie = new Cookie(name, token);
+            cookie.setHttpOnly(true);
+            //배포시에는 아래 주석 풀기
+            //cookie.setSecure(true);
+            cookie.setPath("/");
+            response.addCookie(cookie);
+        }
     }
 
     /**
